Add PSU form factor compatibility and improve component models
Introduces PSU form factor compatibility checks between cases and power supplies, adds SupportedPsuFormFactors to Case, and defines the PsuFormFactor enum. Updates Motherboard with default values for PcieLanes and adds M2Slots. Refines Storage PCIe properties to use int? types. Improves null checks and documentation in CompatibilityEvaluator and RankEvaluator for better clarity and robustness.

+++ Core/Components/Storage.cs
@@ -9,12 +9,12 @@ public class Storage : BaseComponent
     /// <summary>
     /// PcieType, if "Pcie5.0x4" =>> 5.0
     /// </summary>
-    public double? PcieType { get; set; }
+    public int? PcieType { get; set; }
 
     /// <summary>
     /// NVMe pcie lanes needed, if "Pcie5.0x4" >> 4
     /// </summary>
-    public double? PcieLanes { get; set; }
+    public int? PcieLanes { get; set; }
 
     /// <summary>
     /// Read Speed in MB/s
